chore: 주석 삭제 및 타입 분리 #118
