+++ frontend/src/components/common/Sidebar/SidebarButton.tsx
@@ -12,7 +12,6 @@ interface SidebarButtonProps {
 
 const SidebarButton = ({ id, text, icon, icon_focused, size, onClick }: SidebarButtonProps) => {
   const { page } = sidebarStore();
-  // const [isFocused, setIsFocused] = useState(false);
   return (
     <S.SidebarBtn className={`${page === id ? 'active' : ''}`} onClick={onClick}>
       <img alt="icon" src={page === id ? icon_focused : icon} style={{ width: size, height: size }} />
