+++ frontend/src/types/profile.ts
@@ -1,0 +1,7 @@
+export interface ProfileType {
+    officeId: number;
+    name: string;
+    officeName: string;
+    profileUrl: string;
+    rank: string;
+  }
